I don't like this fix
